added sp and bp opcodes for functions

+++ symbols.h
@@ -5,16 +5,16 @@
 #include <unordered_map>
 class Symbols {
     public:
-        long scope_level;
-        long local_offset;
-        long param_offset;
-        long symbols_count;
+        int32_t scope_level;
+        int32_t local_offset;
+        int32_t param_offset;
+        int32_t symbols_count;
         Symbols *next_scope;
         Symbols *prev_scope;
-        std::unordered_map<std::string, long> offsets{};
+        std::unordered_map<std::string, int32_t> offsets{};
         std::unordered_map<std::string, size_t> sizes{};
         Symbols (Symbols *prev_scope, size_t local_offset, long scope_level);
-        int get_stack_offset (char *variable_name, size_t len);
+        int32_t get_stack_offset (char *variable_name, size_t len);
         bool declare_local_variable (char *variable_name, size_t len);
         bool declare_function_parameter (char *parameter_name, size_t len);
         bool has_symbol (char *symbol, size_t len);
